fix(iouring): widen the conn generation to 32 bits (#470)
Follow-up to the process-monotonic generation, from an adversarial review of
that fix. Two of its claims were wrong and the comment I wrote asserted them:

1. The wrap is PROCESS-WIDE, not per-fd. connGenSeq is drawn once per accepted
   fd across every worker, so a wrap costs 65536 accepts ANYWHERE. At this
   cluster's measured end-to-end rate (1,437,649 conns / 45s = ~31.9k
   accepts/s) a 16-bit generation wrapped every 2.05 SECONDS.

2. The binding window is NOT the microsecond cancel latency. It is the paths
   where no cancel is submitted at all: cancelConnOps skips the ASYNC_CANCEL
   when the SQ ring is full (worker.go:2504 has no else branch), leaving the
   op to the 5s pendingRelease backstop; and an armed udHeaderTimer lives for
   ReadHeaderTimeout, 10s by default. Those windows spanned 2.4 and 4.9 wraps
   of a 16-bit generation.

So the 16-bit fix reduced the #470 misroute by 3-5 orders of magnitude
(consistent with 0 hangs in 348,192 connections) but did not close it. The
residual per stranded op was ~1.3e-3 at the validator's C=120 shape.

The generation field was 16 bits only because the fd field was given 40 --
and the layout comment already noted fds and fixed-file indices are "both
< 65536 in practice". onAcceptedFD in fact REJECTS any fd >= fixedFileTableSize
(65536) outright, so 24 bits of fd is 256x the engine's own hard bound.

Re-splits user_data as op(56-63) | gen(24-55, 32 bits) | fd(0-23, 24 bits).
A collision now needs 2^32 intervening accepts -- ~37 hours at 31.9k/s -- which
no engine-side window can span. Same shift-or on the hot path: no added logic,
no measurable cost.

Also corrects the over-claiming comment in acquireConnState, and strengthens
the encoding tests: gens now span the full 32-bit range including the old
16-bit ceiling, and TestGenerationDiffersAcrossReuse asserts uniqueness
UNCONDITIONALLY. Its previous version could not -- it noted that "a fresh pool
object starts at generation 1, colliding with a once-used object's 1" and
worked around it. That comment was describing celeris#470 as a test flake.

+++ engine/iouring/conn_generation_test.go
@@ -26,7 +26,7 @@ import (
 // object is freshly allocated. Pre-fix this fails on the second iteration.
 func TestConnGenerationUniqueAcrossFreshAllocations(t *testing.T) {
 	const n = 256
-	seen := make(map[uint16]int, n)
+	seen := make(map[uint32]int, n)
 	held := make([]*connState, 0, n) // never released: forces fresh allocations
 	for i := 0; i < n; i++ {
 		cs := acquireConnState(context.Background(), 42, 0, false)
@@ -65,7 +65,7 @@ func TestConnGenerationNeverZero(t *testing.T) {
 // too: even when the pool DOES hand back the same object, consecutive
 // occupants of an fd must not share a generation.
 func TestConnGenerationDistinctAcrossPoolRecycle(t *testing.T) {
-	seen := make(map[uint16]bool, 1024)
+	seen := make(map[uint32]bool, 1024)
 	for i := 0; i < 1024; i++ {
 		cs := acquireConnState(context.Background(), 99, 0, false)
 		if seen[cs.generation] {
